refactor: update comments folder structure
