+++ src/discussions/common/HoverCard.test.jsx
@@ -13,14 +13,14 @@ import { AppProvider } from '@edx/frontend-platform/react';
 
 import { initializeStore } from '../../store';
 import { executeThunk } from '../../test-utils';
-import { getCommentsApiUrl } from '../comments/data/api';
 import DiscussionContent from '../discussions-home/DiscussionContent';
+import { getCommentsApiUrl } from '../post-comments/data/api';
 import { getThreadsApiUrl } from '../posts/data/api';
 import { fetchThreads } from '../posts/data/thunks';
 import { DiscussionContext } from './context';
 
 import '../posts/data/__factories__';
-import '../comments/data/__factories__';
+import '../post-comments/data/__factories__';
 
 const commentsApiUrl = getCommentsApiUrl();
 const threadsApiUrl = getThreadsApiUrl();
