+++ src/discussions/index.js
@@ -1,4 +1,4 @@
-export * from './comments';
 export * from './discussions-home';
+export * from './post-comments';
 export * from './posts';
 export * from './topics';
