+++ src/discussions/post-comments/comments/comment/Comment.jsx
@@ -7,18 +7,18 @@ import { useDispatch, useSelector } from 'react-redux';
 import { injectIntl, intlShape } from '@edx/frontend-platform/i18n';
 import { Button, useToggle } from '@edx/paragon';
 
-import HTMLLoader from '../../../components/HTMLLoader';
-import { ContentActions, EndorsementStatus } from '../../../data/constants';
-import { AlertBanner, Confirmation, EndorsedAlertBanner } from '../../common';
-import { DiscussionContext } from '../../common/context';
-import HoverCard from '../../common/HoverCard';
-import { useUserCanAddThreadInBlackoutDate } from '../../data/hooks';
-import { fetchThread } from '../../posts/data/thunks';
-import LikeButton from '../../posts/post/LikeButton';
-import { useActions } from '../../utils';
-import { selectCommentCurrentPage, selectCommentHasMorePages, selectCommentResponses } from '../data/selectors';
-import { editComment, fetchCommentResponses, removeComment } from '../data/thunks';
-import messages from '../messages';
+import HTMLLoader from '../../../../components/HTMLLoader';
+import { ContentActions, EndorsementStatus } from '../../../../data/constants';
+import { AlertBanner, Confirmation, EndorsedAlertBanner } from '../../../common';
+import { DiscussionContext } from '../../../common/context';
+import HoverCard from '../../../common/HoverCard';
+import { useUserCanAddThreadInBlackoutDate } from '../../../data/hooks';
+import { fetchThread } from '../../../posts/data/thunks';
+import LikeButton from '../../../posts/post/LikeButton';
+import { useActions } from '../../../utils';
+import { selectCommentCurrentPage, selectCommentHasMorePages, selectCommentResponses } from '../../data/selectors';
+import { editComment, fetchCommentResponses, removeComment } from '../../data/thunks';
+import messages from '../../messages';
 import CommentEditor from './CommentEditor';
 import CommentHeader from './CommentHeader';
 import { commentShape } from './proptypes';
