+++ src/discussions/post-comments/data/slices.js
@@ -57,15 +57,6 @@ const commentsSlice = createSlice({
         hasMorePages: Boolean(payload.pagination.next),
       };
       state.commentsById = { ...state.commentsById, ...payload.commentsById };
-      // We sort the comments by creation time.
-      // This way our new comments are pushed down to the correct
-      // position when more pages of older comments are loaded.
-      state.commentsInThreads[threadId][endorsed].sort(
-        (a, b) => (
-          Date.parse(state.commentsById[a].createdAt)
-          - Date.parse(state.commentsById[b].createdAt)
-        ),
-      );
     },
     fetchCommentsFailed: (state) => {
       state.status = RequestStatus.FAILED;
@@ -91,12 +82,6 @@ const commentsSlice = createSlice({
         ]),
       ];
       state.commentsById = { ...state.commentsById, ...payload.commentsById };
-      state.commentsInComments[payload.commentId].sort(
-        (a, b) => (
-          Date.parse(state.commentsById[a].createdAt)
-          - Date.parse(state.commentsById[b].createdAt)
-        ),
-      );
       state.responsesPagination[payload.commentId] = {
         currentPage: payload.page,
         totalPages: payload.pagination.numPages,
