+++ src/discussions/post-comments/index.js
@@ -1,0 +1,2 @@
+/* eslint-disable import/prefer-default-export */
+export { default as PostCommentsView } from './PostCommentsView';
