+++ src/index.scss
@@ -478,3 +478,8 @@ header {
 .line-height-24 {
   line-height: 24px;
 }
+
+.comments-sort {
+  margin-bottom: -44px;
+  z-index: 1;
+}
