+++ src/store.js
@@ -3,10 +3,10 @@ import { configureStore } from '@reduxjs/toolkit';
 import { courseTabsReducer } from './components/NavigationBar/data';
 import { blocksReducer } from './data/slices';
 import { cohortsReducer } from './discussions/cohorts/data';
-import { commentsReducer } from './discussions/comments/data';
 import { configReducer } from './discussions/data/slices';
 import { inContextTopicsReducer } from './discussions/in-context-topics/data';
 import { learnersReducer } from './discussions/learners/data';
+import { commentsReducer } from './discussions/post-comments/data';
 import { threadsReducer } from './discussions/posts/data';
 import { topicsReducer } from './discussions/topics/data';
 import { toursReducer } from './discussions/tours/data';
